move sympy2gem to a separate module
