+++ finat/fiat_elements.py
@@ -1,11 +1,8 @@
 from __future__ import absolute_import, print_function, division
 from six import iteritems
 
-from functools import reduce
-
 import numpy as np
 import sympy as sp
-from singledispatch import singledispatch
 
 import FIAT
 from FIAT.polynomial_set import mis, form_matrix_product
@@ -13,6 +10,7 @@ from FIAT.polynomial_set import mis, form_matrix_product
 import gem
 
 from finat.finiteelementbase import FiniteElementBase
+from finat.sympy2gem import sympy2gem
 
 
 class FiatElementBase(FiniteElementBase):
@@ -229,45 +227,3 @@ class Nedelec(VectorFiatElement):
 class NedelecSecondKind(VectorFiatElement):
     def __init__(self, cell, degree):
         super(NedelecSecondKind, self).__init__(FIAT.NedelecSecondKind(cell, degree))
-
-
-@singledispatch
-def sympy2gem(node, self):
-    raise AssertionError("sympy node expected, got %s" % type(node))
-
-
-@sympy2gem.register(sp.Expr)
-def sympy2gem_expr(node, self):
-    raise NotImplementedError("no handler for sympy node type %s" % type(node))
-
-
-@sympy2gem.register(sp.Add)
-def sympy2gem_add(node, self):
-    return reduce(gem.Sum, map(self, node.args))
-
-
-@sympy2gem.register(sp.Mul)
-def sympy2gem_mul(node, self):
-    return reduce(gem.Product, map(self, node.args))
-
-
-@sympy2gem.register(sp.Pow)
-def sympy2gem_pow(node, self):
-    return gem.Power(*map(self, node.args))
-
-
-@sympy2gem.register(sp.Integer)
-@sympy2gem.register(int)
-def sympy2gem_integer(node, self):
-    return gem.Literal(node)
-
-
-@sympy2gem.register(sp.Float)
-@sympy2gem.register(float)
-def sympy2gem_float(node, self):
-    return gem.Literal(node)
-
-
-@sympy2gem.register(sp.Symbol)
-def sympy2gem_symbol(node, self):
-    return self.bindings[node]
